stack vagrant

+++ provisioners/shell/plateform/installer-sfynx-nginx.sh
@@ -5,14 +5,15 @@ PLATEFORM_INSTALL_TYPE=$3
 PLATEFORM_INSTALL_VERSION=$4
 PLATEFORM_PROJET_NAME=$5
 PLATEFORM_PROJET_GIT=$6
+INSTALL_USERWWW=$7
 source $DIR/provisioners/shell/env.sh
 
 #if var is empty
 if [ -z "$PLATEFORM_PROJET_GIT" ]; then
-    $PLATEFORM_PROJET_GIT="https://github.com/RappFrance/rapp_nosbelidees"
+    $PLATEFORM_PROJET_GIT="https://github.com/pigroupe/cmf-sfynx.git"
 fi
 
-# we create directories
+echo "**** we create directories ****"
 if [ ! -d $INSTALL_USERWWW ]; then
     mkdir -p $INSTALL_USERWWW
 fi
@@ -20,34 +21,43 @@ cd $INSTALL_USERWWW
 
 # we create project
 if [ ! -d $PLATEFORM_PROJET_NAME ]; then
-    #git clone https://github.com/pigroupe/cmf-sfynx.git $PLATEFORM_PROJET_NAME
+    #git clone $PLATEFORM_PROJET_GIT $PLATEFORM_PROJET_NAME
     mkdir -p $PLATEFORM_PROJET_NAME
 fi
 cd $PLATEFORM_PROJET_NAME
 
-# we create default directories
-mkdir -p app/cache
-mkdir -p app/logs
-mkdir -p app/cachesfynx/loginfailure
-mkdir -p web/uploads/media
-mkdir -p web/yui
-rm app/config/parameters.yml
-cp app/config/parameters.yml.dist app/config/parameters.yml
+echo "**** we create default directories ****"
+if [ ! -d app/cachesfynx ]; then
+    mkdir -p app/cache
+    mkdir -p app/logs
+    mkdir -p app/cachesfynx/loginfailure
+    mkdir -p web/uploads/media
+    mkdir -p web/yui
+fi
+if [ ! -f app/config/parameters.yml ]; then
+    cp app/config/parameters.yml.dist app/config/parameters.yml
+    sed -i 's/%%/%/g' app/config/parameters.yml
+fi
+if [ ! -f app/phpunit.xml ]; then
+    cp app/phpunit.xml.dist app/phpunit.xml
+fi
 
-# we add env var
+echo "**** we add env variables ****"
+if ! grep -q "SYMFONY__DATABASE__NAME__ENV" ~/.profile; then
 cat <<EOT >> ~/.profile
 
 # env vars for SFYNX platform
-export SYMFONY__DATABASE__NAME__ENV=sfynx$PLATEFORM_PROJET_NAME_dev;
+export SYMFONY__DATABASE__NAME__ENV=sfynx_$PLATEFORM_PROJET_NAME_dev;
 export SYMFONY__DATABASE__USER__ENV=root;
 export SYMFONY__DATABASE__PASSWORD__ENV=pacman;
-export SYMFONY__TEST__DATABASE__NAME__ENV=sfynx$PLATEFORM_PROJET_NAME_test;
+export SYMFONY__TEST__DATABASE__NAME__ENV=sfynx_$PLATEFORM_PROJET_NAME_test;
 export SYMFONY__TEST__DATABASE__USER__ENV=root;
 export SYMFONY__TEST__DATABASE__PASSWORD__ENV=pacman;
 EOT
 source ~/.profile
+fi
 
-# we create the virtualhiost of sfynx for nginx
+echo "**** we create the virtualhost ****"
 mkdir -p /tmp
 cat <<EOT >/tmp/$PLATEFORM_PROJET_NAME
 #upstream php5-fpm-sock {  
@@ -133,6 +143,12 @@ server {
         fastcgi_param SCRIPT_FILENAME \$document_root\$fastcgi_script_name;
         fastcgi_param  HTTPS off;
         # fastcgi_param PHP_VALUE "auto_prepend_file=$INSTALL_USERWWW/xhprof/external/header.php \n auto_append_file=$INSTALL_USERWWW/xhprof/external/footer.php";
+        fastcgi_param SYMFONY__DATABASE__NAME__ENV sfynx_$PLATEFORM_PROJET_NAME_dev;
+        fastcgi_param SYMFONY__DATABASE__USER__ENV root;
+        fastcgi_param SYMFONY__DATABASE__PASSWORD__ENV pacman;
+        fastcgi_param SYMFONY__TEST__DATABASE__NAME__ENV sfynx_$PLATEFORM_PROJET_NAME_test;
+        fastcgi_param SYMFONY__TEST__DATABASE__USER__ENV root;
+        fastcgi_param SYMFONY__TEST__DATABASE__PASSWORD__ENV pacman;
     }
 
     # Nginx Cache Control for Static Files
@@ -251,6 +267,12 @@ server {
         fastcgi_param SCRIPT_FILENAME \$document_root\$fastcgi_script_name;
         fastcgi_param  HTTPS off;
         # fastcgi_param PHP_VALUE "auto_prepend_file=$INSTALL_USERWWW/xhprof/external/header.php \n auto_append_file=$INSTALL_USERWWW/xhprof/external/footer.php";
+        fastcgi_param SYMFONY__DATABASE__NAME__ENV sfynx_$PLATEFORM_PROJET_NAME_dev;
+        fastcgi_param SYMFONY__DATABASE__USER__ENV root;
+        fastcgi_param SYMFONY__DATABASE__PASSWORD__ENV pacman;
+        fastcgi_param SYMFONY__TEST__DATABASE__NAME__ENV sfynx_$PLATEFORM_PROJET_NAME_test;
+        fastcgi_param SYMFONY__TEST__DATABASE__USER__ENV root;
+        fastcgi_param SYMFONY__TEST__DATABASE__PASSWORD__ENV pacman;
     }
 
     # Nginx Cache Control for Static Files
@@ -369,6 +391,12 @@ server {
         fastcgi_param SCRIPT_FILENAME \$document_root\$fastcgi_script_name;
         fastcgi_param  HTTPS off;
         # fastcgi_param PHP_VALUE "auto_prepend_file=$INSTALL_USERWWW/xhprof/external/header.php \n auto_append_file=$INSTALL_USERWWW/xhprof/external/footer.php";
+        fastcgi_param SYMFONY__DATABASE__NAME__ENV sfynx_$PLATEFORM_PROJET_NAME_dev;
+        fastcgi_param SYMFONY__DATABASE__USER__ENV root;
+        fastcgi_param SYMFONY__DATABASE__PASSWORD__ENV pacman;
+        fastcgi_param SYMFONY__TEST__DATABASE__NAME__ENV sfynx_$PLATEFORM_PROJET_NAME_test;
+        fastcgi_param SYMFONY__TEST__DATABASE__USER__ENV root;
+        fastcgi_param SYMFONY__TEST__DATABASE__PASSWORD__ENV pacman;
     }
 
     # Nginx Cache Control for Static Files
@@ -408,10 +436,10 @@ server {
 
 }
 EOT
-mv /tmp/$PLATEFORM_PROJET_NAME /etc/nginx/sites-available/$PLATEFORM_PROJET_NAME
+sudo mv /tmp/$PLATEFORM_PROJET_NAME /etc/nginx/sites-available/$PLATEFORM_PROJET_NAME
 
 # we create the symbilic link
-ln -s /etc/nginx/sites-available/$PLATEFORM_PROJET_NAME /etc/nginx/sites-enabled/$PLATEFORM_PROJET_NAME
+sudo ln -s /etc/nginx/sites-available/$PLATEFORM_PROJET_NAME /etc/nginx/sites-enabled/$PLATEFORM_PROJET_NAME
 
 # we add host in the /etc/hosts file
 if ! grep -q "dev.$PLATEFORM_PROJET_NAME.local" /etc/hosts; then
@@ -421,18 +449,59 @@ if ! grep -q "dev.$PLATEFORM_PROJET_NAME.local" /etc/hosts; then
     echo "127.0.0.1    prod.$PLATEFORM_PROJET_NAME.local" | tee --append /etc/hosts
 fi
 
-# we restart nginx server
+echo "**** we restart nginx server ****"
 sudo service nginx restart
 
-# we install the composer file
+echo "**** we install/update the composer file ****"
 if [ ! -f composer.phar ]; then
     wget https://getcomposer.org/composer.phar -O ./composer.phar
-    # curl -s https://getcomposer.org/installer | php
+else
+    php composer.phar self-update
 fi
+echo "**** we lauch the composer ****"
 php -d memory_limit=1024M composer.phar install --no-interaction
+echo "**** Generating optimized autoload files ****"
 php composer.phar dump-autoload --optimize
 
-# create database
+echo "**** we remove cache files ****"
+rm -rf app/cache/*
+rm -rf app/logs/*
+
+echo "**** we set all necessary permissions ****"
+# Utiliser l'ACL sur un système qui supporte chmod +a
+#HTTPDUSER=`ps aux | grep -E '[a]pache|[h]ttpd|[_]www|[w]ww-data|[n]ginx' | grep -v root | head -1 | cut -d\  -f1`
+#sudo chmod +a "$HTTPDUSER allow delete,write,append,file_inherit,directory_inherit" app/cache app/logs
+#sudo chmod +a "`whoami` allow delete,write,append,file_inherit,directory_inherit" app/cache app/logs
+
+# Utiliser l'ACL sur un système qui ne supporte pas chmod +a
+HTTPDUSER=`ps aux | grep -E '[a]pache|[h]ttpd|[_]www|[w]ww-data|[n]ginx' | grep -v root | head -1 | cut -d\  -f1`
+sudo setfacl -R -m u:"$HTTPDUSER":rwX -m u:`whoami`:rwX app/cache app/logs
+sudo setfacl -dR -m u:"$HTTPDUSER":rwX -m u:`whoami`:rwX app/cache app/logs
+
+# sans utiliser ACL
+## Définit une permission 0775 aux fichiers
+#echo "umask(0002);" | sudo tee --prepend app/console
+#echo "umask(0002);" | sudo tee --prepend web/app_dev.php
+#echo "umask(0002);" | sudo tee --prepend web/app.php
+## Définit une permission 0777 aux fichiers
+#echo "umask(0000);" | sudo tee --prepend app/console
+#echo "umask(0000);" | sudo tee --prepend web/app_dev.php
+#echo "umask(0000);" | sudo tee --prepend web/app.php
+
+
+
+# permission
+#sudo chown -R root:www-data app/cache
+#sudo chown -R root:www-data app/logs
+#sudo chown -R root:www-data app/config/parameters.yml
+#sudo chown -R root:www-data web/uploads
+#sudo chmod -R 775 app/config/parameters.yml
+#sudo chmod -R 775 app/cache
+#sudo chmod -R 775 app/logs
+#sudo chmod -R 775 web/uploads
+#sudo chown -R www-data:www-data $INSTALL_USERWWW/$PLATEFORM_PROJET_NAME
+
+echo "**** we create database ****"
 php app/console doctrine:database:create
 php app/console doctrine:schema:create
 php app/console doctrine:fixtures:load
@@ -440,5 +509,5 @@ php app/console assets:install
 php app/console assetic:dump
 php app/console clear:cache
 
-# we run the phing script to initialize the sfynx project
+echo "**** we run the phing script to initialize the project ****"
 vendor/bin/phing -f app/phing/initialize.xml rebuild
